requirements.txt
pip>=21.3.1
ipaddress>=1.0.23
setuptools>=44.1.1
ansible-core>=2.16.0
requests>=2.31.0
ntnx-clustermgmt-py-client<=4.1.1
ntnx-iam-py-client<=4.0.1
ntnx-microseg-py-client<=4.1.1
ntnx-networking-py-client<=4.1.1
ntnx-prism-py-client<=4.1.1
ntnx-vmm-py-client<=4.1.1
ntnx-volumes-py-client<=4.1.1
ntnx-dataprotection-py-client<=4.1.1
ntnx_datapolicies_py_client<=4.1.1
ntnx_lifecycle_py_client<=4.1.1
ntnx-objects-py-client<=4.0.1
ntnx-security-py-client<=4.0.1
ntnx-licensing-py-client<=4.2.1
